requirements.txt
streamlit==1.40.0
pandas==2.2.0
bcrypt==4.1.2
plotly==5.18.0
streamlit-plotly-events==0.0.6
paho-mqtt>=2.0
